Moved xdoc extensions into own file
