+++ OmniSharp/ProjectManipulation/RemoveFromProject/RemoveFromProjectHandler.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
@@ -30,9 +29,7 @@ namespace OmniSharp.ProjectManipulation.RemoveFromProject
             var relativeFileName = request.FileName.Replace(relativeProject.FileName.Substring(0, relativeProject.FileName.LastIndexOf(Path.DirectorySeparatorChar) + 1), "")
                 .Replace("/", @"\");
 
-            var compilationNodes = project.CompilationNodes();
-
-            var fileNode = compilationNodes.FirstOrDefault(n => n.Attribute("Include").Value.Equals(relativeFileName, StringComparison.InvariantCultureIgnoreCase));
+            var fileNode = project.CompilationNodes().FirstOrDefault(n => n.Attribute("Include").Value.Equals(relativeFileName, StringComparison.InvariantCultureIgnoreCase));
 
             if (fileNode != null)
             {
@@ -54,24 +51,6 @@ namespace OmniSharp.ProjectManipulation.RemoveFromProject
         private void RemoveFileFromProject(IProject project, string filename)
         {
             project.Files.Remove(project.Files.First(f => f.FileName == filename));
-        }
-    }
-
-    public static class XDocumentExtensions
-    {
-        private static readonly XNamespace _msBuildNameSpace = "http://schemas.microsoft.com/developer/msbuild/2003";
-
-        public static IEnumerable<XElement> ItemGroupNodes(this XDocument document)
-        {
-            return document.Element(_msBuildNameSpace + "Project")
-                          .Elements(_msBuildNameSpace + "ItemGroup");
-        }
-
-        public static IEnumerable<XElement> CompilationNodes(this XDocument document)
-        {
-            return document.Element(_msBuildNameSpace + "Project")
-                          .Elements(_msBuildNameSpace + "ItemGroup")
-                          .Elements(_msBuildNameSpace + "Compile");
         }
     }
 }
